+++ avk/algorithms/bitonic_sort.cpp
@@ -2,13 +2,13 @@
 
 void bitonic_sort(main_array array)
 {
-	for (uint k = 2; k <= array.size(); k *= 2)
+	for (size_t k = 2; k <= array.size(); k *= 2)
 	{
-		for (uint j = k / 2; j > 0; j /= 2)
+		for (size_t j = k / 2; j > 0; j /= 2)
 		{
-			for (uint i = 0; i < array.size(); ++i)
+			for (size_t i = 0; i < array.size(); ++i)
 			{
-				uint l = i ^ j;
+				size_t l = i ^ j;
 				if (l > i)
 				{
 					if (((i & k) == 0 && (array[i] > array[l])) ||
