+++ avk/algorithms/bose_nelson_network.cpp
@@ -1,21 +1,21 @@
 #include "all.h"
 
 
-static void ranged_cas(main_array array, uint a, uint b, uint offset)
+static void ranged_cas(main_array array, size_t a, size_t b, size_t offset)
 {
-	int half = (b - a) / 2, m = a + half;
+	size_t half = (b - a) / 2, m = a + half;
 	a += offset;
 
-	for (uint i = 0; i < half - offset; i++)
+	for (size_t i = 0; i < half - offset; i++)
 		if ((i & ~offset) == i && m + i < array.size())
 			compare_swap(array, a + i, m + i);
 }
 
 void bose_nelson_network(main_array array)
 {
-	uint current_size = 1 << (uint)(ceilf(logf((float)array.size()) / logf(2))); //@TODO AAAAH
-	for (uint k = 2; k <= current_size; k *= 2)
-		for (uint j = 0; j < k / 2; j++)
-			for (uint i = 0; i + j < array.size(); i += k)
+	size_t current_size = (size_t)1 << (size_t)(ceilf(logf((float)array.size()) / logf(2))); //@TODO AAAAH
+	for (size_t k = 2; k <= current_size; k *= 2)
+		for (size_t j = 0; j < k / 2; j++)
+			for (size_t i = 0; i + j < array.size(); i += k)
 				ranged_cas(array, i, i + k, j);
 }
