+++ avk/algorithms/bose_nelson_network_parallel.cpp
@@ -1,6 +1,6 @@
 #include "all.h"
 
-static void bose_nelson_merge(main_array array, uint start1, uint len1, uint start2, uint len2)
+static void bose_nelson_merge(main_array array, size_t start1, size_t len1, size_t start2, size_t len2)
 {
     if (len1 == 1 && len2 == 1)
     {
@@ -18,36 +18,36 @@ static void bose_nelson_merge(main_array array, uint start1, uint len1, uint sta
     }
     else
     {
-        uint mid1 = len1 / 2;
-        uint mid2 = len1 % 2 == 1 ? len2 / 2 : (len2 + 1) / 2;
-        uint params[3][4] =
+        size_t mid1 = len1 / 2;
+        size_t mid2 = len1 % 2 == 1 ? len2 / 2 : (len2 + 1) / 2;
+        size_t params[3][4] =
         {
             { start1 , mid1 , start2 , mid2 },
             { start1 + mid1 , len1 - mid1 , start2 + mid2 , len2 - mid2 },
             { start1 + mid1 , len1 - mid1 , start2 , mid2 }
         };
 
-        parallel_for<uint>(0, 3, [&](uint i)
+        parallel_for<size_t>(0, 3, [&](size_t i)
         {
             bose_nelson_merge(array, params[i][0], params[i][1], params[i][2], params[i][3]);
         });
     }
 }
 
-static void bose_nelson_core(main_array array, uint start, uint length)
+static void bose_nelson_core(main_array array, size_t start, size_t length)
 {
     if (length < 2)
         return;
 
-    uint mid = length / 2;
+    size_t mid = length / 2;
 
-    uint params[2][2] =
+    size_t params[2][2] =
     {
         { start, mid },
         { start + mid, length - mid }
     };
     
-    parallel_for<uint>(0, 2, [&](uint i)
+    parallel_for<size_t>(0, 2, [&](size_t i)
     {
         bose_nelson_core(array, params[i][0], params[i][1]);
     });
