+++ avk/algorithms/counting_sort.cpp
@@ -3,12 +3,12 @@
 
 void counting_sort(main_array array)
 {
-	std::vector<uint> counts;
+	std::vector<size_t> counts;
 	std::vector<item> buffer;
 	buffer.resize(array.size());
 
-	uint max = 0;
-	for (uint i = 0; i < array.size(); ++i)
+	size_t max = 0;
+	for (size_t i = 0; i < array.size(); ++i)
 	{
 		if (array[i].value > max)
 			max = array[i].value;
@@ -20,16 +20,16 @@ void counting_sort(main_array array)
 	counts.resize(max + 1);
 	for (item& e : buffer)
 		++counts[e.value];
-	uint offset = 0;
-	for (uint& e : counts)
+	size_t offset = 0;
+	for (size_t& e : counts)
 	{
-		uint k = e;
+		size_t k = e;
 		e = offset;
 		offset += k;
 	}
 	for (item& e : buffer)
 	{
-		uint& k = counts[e.value];
+		size_t& k = counts[e.value];
 		array[k] = e;
 		++k;
 	}
