+++ avk/algorithms/fold_sort_bottom_up.cpp
@@ -1,7 +1,7 @@
 #include "all.h"
 #include "sort_utility.h"
 
-static void halver(main_array array, uint low, uint high)
+static void halver(main_array array, size_t low, size_t high)
 {
 	--high;
 	while (low < high)
@@ -14,14 +14,14 @@ static void halver(main_array array, uint low, uint high)
 
 void fold_sort_bottom_up(main_array array)
 {
-	uint size = array.size();
-	uint log2 = fast_log2(size);
+	size_t size = array.size();
+	size_t log2 = fast_log2(size);
 
-	//for (uint n = 0; n <= log2; ++n)
+	//for (size_t n = 0; n <= log2; ++n)
 	{
-		for (uint step = 1; step <= size; step *= 2)
+		for (size_t step = 1; step <= size; step *= 2)
 		{
-			for (uint i = 0; i < size; i += step)
+			for (size_t i = 0; i < size; i += step)
 			{
 				halver(array, i, i + step);
 			}
