+++ avk/algorithms/grail_sort.cpp
@@ -3,7 +3,7 @@
 
 static int gs_cmp(item* l, item* r)
 {
-	return compare(*l, *r);
+	return (int)compare(*l, *r);
 }
 #define SORT_CMP gs_cmp
 #define SORT_TYPE item
@@ -16,10 +16,10 @@ void grail_sort(main_array array)
 	{
 		std::vector<item> buffer;
 		buffer.resize(sort_config::grail_sort_buffer_size);
-		grail_commonSort(array.data(), array.size(), buffer.data(), (int)buffer.size());
+		grail_commonSort(array.data(), (int)array.size(), buffer.data(), (int)buffer.size());
 	}
 	else
 	{
-		GrailSort(array.data(), array.size());
+		GrailSort(array.data(), (int)array.size());
 	}
 }
