+++ avk/algorithms/odd_even_merge_sort_parallel.cpp
@@ -1,32 +1,32 @@
 #include "all.h"
 
-static void odd_even_merge_sort_kernel(main_array array, uint lo, uint m2, uint n, uint r)
+static void odd_even_merge_sort_kernel(main_array array, size_t lo, size_t m2, size_t n, size_t r)
 {
-    int m = r * 2;
+    size_t m = r * 2;
     if (m < n)
     {
         if ((n / r) % 2 != 0)
         {
-            uint params[2][3] =
+            size_t params[2][3] =
             {
                 { lo, (m2 + 1) / 2, n + r },
                 { lo + r, m2 / 2, n - r }
             };
 
-            parallel_for<uint>(0, 2, [=](uint i)
+            parallel_for<size_t>(0, 2, [=](size_t i)
             {
                 odd_even_merge_sort_kernel(array, params[i][0], params[i][1], params[i][2], m);
             });
         }
         else
         {
-            uint params[2][3] =
+            size_t params[2][3] =
             {
                 { lo, (m2 + 1) / 2, n },
                 { lo + r, m2 / 2, n }
             };
 
-            parallel_for<uint>(0, 2, [=](uint i)
+            parallel_for<size_t>(0, 2, [=](size_t i)
             {
                 odd_even_merge_sort_kernel(array, params[i][0], params[i][1], params[i][2], m);
             });
@@ -34,14 +34,14 @@ static void odd_even_merge_sort_kernel(main_array array, uint lo, uint m2, uint 
 
         if (m2 % 2 != 0)
         {
-            for (int i = lo; i + r < lo + n; i += m)
+            for (size_t i = lo; i + r < lo + n; i += m)
             {
                 compare_swap(array, i, i + r);
             }
         }
         else
         {
-            for (int i = lo + r; i + r < lo + n; i += m)
+            for (size_t i = lo + r; i + r < lo + n; i += m)
             {
                 compare_swap(array, i, i + r);
             }
@@ -56,18 +56,18 @@ static void odd_even_merge_sort_kernel(main_array array, uint lo, uint m2, uint 
     }
 }
 
-static void odd_even_merge_sort_core(main_array array, uint lo, uint n)
+static void odd_even_merge_sort_core(main_array array, size_t lo, size_t n)
 {
     if (n > 1)
     {
-        int m = n / 2;
-        uint params[2][2] =
+        size_t m = n / 2;
+        size_t params[2][2] =
         {
             { lo, m },
             { lo + m, n - m }
         };
         
-        parallel_for<uint>(0, 2, [=](uint i)
+        parallel_for<size_t>(0, 2, [=](size_t i)
         {
             odd_even_merge_sort_core(array, params[i][0], params[i][1]);
         });
