+++ avk/algorithms/sqrt_sort.cpp
@@ -2,7 +2,7 @@
 
 static int ss_cmp(item* l, item* r)
 {
-	return compare(*l, *r);
+	return (int)compare(*l, *r);
 }
 #define SORT_CMP ss_cmp
 #define SORT_TYPE item
@@ -10,5 +10,5 @@ static int ss_cmp(item* l, item* r)
 
 void sqrt_sort(main_array array)
 {
-	SqrtSort(array.data(), array.size());
+	SqrtSort(array.data(), (int)array.size());
 }
